Added checks and tests.

+++ BasketApi/BasketApi/Controllers/ItemsController.cs
@@ -2,7 +2,6 @@
 using BasketApi.Representations;
 using Microsoft.AspNetCore.Mvc;
 using System;
-using System.Collections.Generic;
 
 namespace BasketApi.Controllers
 {
@@ -16,16 +15,18 @@ namespace BasketApi.Controllers
             this.basketRepository = basketRepository;
         }
 
-        [HttpGet]
-        public IEnumerable<Item> GetItemsForBasket(Guid basketId)
-        {
-            return this.basketRepository.GetItemsForBasket(basketId);
-        }
-
         [HttpGet("{itemCode}")]
         public IActionResult Get(Guid basketId, string itemCode)
         {
+            var basket = this.basketRepository.GetBasket(basketId);
+
+            if (basket == null)
+                return NotFound();
+            
             var item = this.basketRepository.GetItem(basketId, itemCode);
+
+            if (item == null)
+                return NotFound();
 
             return new OkObjectResult(item);
         }
@@ -33,6 +34,14 @@ namespace BasketApi.Controllers
         [HttpPut]
         public IActionResult Put(Guid basketId, [FromBody] Item item)
         {
+            if (item == null)
+                return BadRequest();
+
+            var basket = this.basketRepository.GetBasket(basketId);
+
+            if (basket == null)
+                return NotFound();
+
             var existingItem = this.basketRepository.GetItem(basketId, item.code);
 
             if (existingItem != null)
@@ -46,9 +55,30 @@ namespace BasketApi.Controllers
         [HttpDelete("{itemCode}")]
         public IActionResult Delete(Guid basketId, string itemCode)
         {
+            var basket = this.basketRepository.GetBasket(basketId);
+
+            if (basket == null)
+                return NotFound();
+
+            var item = this.basketRepository.GetItem(basketId, itemCode);
+
+            if (item == null)
+                return NotFound();
+
             this.basketRepository.RemoveItemFromBasket(basketId, itemCode);
 
             return NoContent();
+        }
+
+        [HttpGet]
+        public IActionResult GetItemsForBasket(Guid basketId)
+        {
+            var basket = this.basketRepository.GetBasket(basketId);
+
+            if (basket == null)
+                return new NotFoundResult();
+
+            return new OkObjectResult(this.basketRepository.GetItemsForBasket(basketId));
         }
     }
 }
